+++ px_user_identity/classes.py
@@ -2,15 +2,9 @@ class UserDeviceAccountRegistration:
     def __init__ (self, public_key, title, location):
         self.public_key: str = public_key
 
-class RequestedOperation:
-    def __init__(self, action: str, security: str, key_type: str, force_operation: bool = False):
-        self.action = action
-        self.security = security
-        self.key_type = key_type
-        self.force_operation = force_operation
-
 class UserClass:
-    def __init__(self, user_type, is_managed, first_name, last_name, email):
+    def __init__(self, user_id, user_type, is_managed, first_name, last_name, email):
+        self.id = user_id
         self.user_type = user_type
         self.is_managed = is_managed
         self.first_name = first_name