+++ px_user_identity/config.py
@@ -1,8 +1,8 @@
 import yaml
 from pathlib import Path
 
-from .log import Logger
-from .filesystem import Filesystem
+from px_device_identity.log import Logger
+from px_device_identity.filesystem import Filesystem
 
 log = Logger('CONFIG')
 
@@ -17,17 +17,19 @@ def USER_CONFIG_DIR():
     return home_path + config_dir
 
 def get_user_config():
-    log.info('Loading user config from {}.'.format(USER_CONFIG_DIR()))
-    fs = Filesystem(CONFIG_DIR(), 'user.yml', 'r')
+    log.info('=> Loading user config from {}.'.format(USER_CONFIG_DIR()))
+    fs = Filesystem(USER_CONFIG_DIR(), 'user.yml', 'r')
     file = fs.open_file()
     config = yaml.load(file, Loader=yaml.BaseLoader)
     cfg_user = {
         'id': config.get('id'),
+        'first_name': config.get('firstName'),
+        'last_name': config.get('lastName'),
+        'email': config.get('email'),
         'userType': config.get('userType'),
         'keySecurity': config.get('keySecurity'),
         'keyType': config.get('keyType'),
         'isManaged': config.get('isManaged'),
-        'host': config.get('host'),
         'configVersion': config.get('configVersion'),
         'initiatedOn': config.get('initiatedOn')
     }