+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.1]
+### Added
+
+- Now supports all features `px-device-identity` does
+- Init, Signing, Get JKW / JWKS
+- Accepts new flags like `--firstname`, `--lastname`, `--email`
+
+### Removed
+
+- It's no longer possible to pass `--host`. This is read from device config by default.
+
 ## [0.0.0]
 ### Added